requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-groq
langgraph>=1.0.1
langgraph-checkpoint>=1.0.1
langgraph-checkpoint-sqlite>=1.0.1
python-dotenv
psycopg2-binary
sqlalchemy